# hashed dbc import