created a hashed list of frames to speedup the import process of dbc files, in my example i used a dbc file with 65k lines and around 800 frames.

before ...
 ```
 Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    50600    5.286    0.000    8.465    0.000 canmatrix.py:1390(frame_by_id)
 25100584    3.180    0.000    3.180    0.000 canmatrix.py:600(__eq__)
        1    0.452    0.452    9.610    9.610 canmatrix\formats\dbc.py:437(load)
   110409    0.117    0.000    0.117    0.000 {method 'match' of '_sre.SRE_Pattern' objects}
```

after ...
   ```
Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.388    0.388    1.190    1.190 canmatrix\formats\dbc.py:437(load)
      791    0.131    0.000    0.134    0.000 {built-in method builtins.print}
   110409    0.104    0.000    0.104    0.000 {method 'match' of '_sre.SRE_Pattern' objects}
   759393    0.078    0.000    0.078    0.000 {method 'startswith' of 'str' objects}
```

so it reduced the overall time needed to import that specific dbc file from 9 to 1 seconds on an i7-8550U at 3,9 GHz.

We are able to do this, because the Frame Objects are mutable anyway and stored in the db by reference.