@altendky 2019-03-06
@ebroecker, do you think we want to do this as an import-only thing?  Or do you want to explore being able to do this in general?

While it's easy to make things hashable doing so properly often means making the objects immutable.  This is because for dicts etc to work as expected, things that compare equal must hash equal and so the hash must depend on the same things an equality check does.  But the object must also hash consistently otherwise it won't work very well as a dict key.  That's easy enough with attrs classes, just `@attr.s(frozen=True)` and it will be immutable and hashable.  The 'cost' here would be that instead of just modifying a frame's identifier such as `frame.arbitration_id.id = 42` you would `frame.arbitration_id = attr.evolve(frame.arbitration_id, id=42)` (or just make a new one manually, it's a small class `frame.arbitration_id = canmatrix.ArbitrationId(id=42, extended=False)`).