@timschneider 2019-03-06
@altendky @ebroecker 
I did this as an import only thing because I don't know if anyone is using canmatrix in the way of creating Frames data driven and assigning or changing ids afterwards or if canmatrix wants to support this.

How would you handle this in general if the ArbitrationId class is frozen? Then we have to keep track of the id in the frame or make the id of the frame immutable aswell or do i miss a thing?