@altendky 2019-03-08
@timschneider certainly good concerns about how people expect to use this.  I'm not going to suggest we jump to immutable everything...  but @ebroecker is pretty open to change generally.

I think of making the `ArbitrationId` immutable as not a huge step generally.  If it weren't a class and there weren't extended/standard then, it'd be an integer which is immutable anyways.  It seems like it shouldn't be a huge hassle to modify the id of a frame as I showed in my example code above.

But...  I do see your point at the next layer around that.  The dict that presumably we'd create and use for this could use the new hashable `ArbitrationId` as a key but...  the frame could have changed in the meantime and have a new id.  I mean we could make the id a property and so on...  but yeah, creating a system that maintains all relationships accurately is a bit more of a task than shifting your PR here to another part of the code.

I think in general we should consider changes in the direction I proposed but since it's not such a small thing as I originally had in my head, I'll withdraw that proposal from this PR discussion.

---

Any opinions about the class-vs-functions-with-closures comment?
https://github.com/ebroecker/canmatrix/pull/319/files#r263016533