@timschneider 2019-03-11
@altendky i see the benefits of a small class and will rewrite this within the next days.